refactor(ui): make general list suspense component
