+++ src/app/main/landlord/my-clinics/page.tsx
@@ -3,7 +3,7 @@ import Link from "next/link";
 import Button from "@/components/Button";
 import { MdAddBox } from "react-icons/md";
 import ClinicFetcher from "./components/ClinicFetcher";
-import ClinicsListSkeleton from "./components/ClinicsListSkeleton";
+import ClinicsListSkeleton from "@/components/ListSuspenseSkeleton/ListSuspenseSkeleton";
 
 export default async function page() {
   return (
